// src/platform/plugins/shared/embeddable/public/drilldowns/types.d.ts
export type DrilldownActionState = DrilldownState & {
    actionId: string;
};
// ...
export type HasDrilldowns = {
    setDrilldowns: (drilldowns: DrilldownState[]) => void;
    drilldowns$: PublishingSubject<DrilldownActionState[]>;
};